feat: Add ContainsList class with search functionality
This commit adds a new class called ContainsList.cs which contains a list of integers. It includes a search functionality that allows users to check if a specific number is present in the list. This feature will be useful for quickly determining if a number exists in the list without iterating through all the elements.